refactor : address 도메인 분리
